+++ src/main/java/com/carcenter/car_center_api/maintenance/dtos/MaintenanceUpdateRequest.java
@@ -21,5 +21,5 @@ public class MaintenanceUpdateRequest {
 
     private Long mechanicId;
 
-    private Double hoursWorked;
+    private Integer hoursWorked;
 }
